Fix build on architectures where off_t is signed by casting to uintmax_t before comparing with a size_t.

+++ usr.bin/head/head.c
@@ -146,7 +146,7 @@ head_bytes(FILE *fp, off_t cnt)
 	size_t readlen;
 
 	while (cnt) {
-		if (cnt < sizeof(buf))
+		if ((uintmax_t)cnt < sizeof(buf))
 			readlen = cnt;
 		else
 			readlen = sizeof(buf);
